In the criteira.R file I made sure to pull out the criteria that show 0 as an artifact from excel in the applicable criteria data frame.
In the toxics analysis I switched back the data frame for the criteria evaluation to the data wo void so we could see the new methods included.

Not sure what changed with the data summary draft excel file. Learning that changes in xlsx files don't show up in the R studio diff.